+++ RSDKv5/RSDK/Graphics/Sprite.cpp
@@ -906,7 +906,7 @@ bool32 RSDK::ImageTGA::Load(const char *fileName, bool32 loadHeader)
 uint16 RSDK::LoadSpriteSheet(const char *filename, int32 scope)
 {
     char fullFilePath[0x100];
-    sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Sprites/%s", filename);
+    sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Sprites/%s", filename);
 
     RETRO_HASH_MD5(hash);
     GEN_HASH_MD5(filename, hash);
@@ -969,7 +969,7 @@ uint16 RSDK::LoadSpriteSheet(const char *filename, int32 scope)
 bool32 RSDK::LoadImage(const char *filename, double displayLength, double fadeSpeed, bool32 (*skipCallback)())
 {
     char fullFilePath[0x100];
-    sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Images/%s", filename);
+    sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Images/%s", filename);
 
 #if RETRO_REV02
     ImagePNG image;
