+++ RSDKv5/RSDK/Storage/Text.cpp
@@ -389,7 +389,7 @@ void RSDK::InitStringList(String *stringList, int32 size)
 void RSDK::LoadStringList(String *stringList, const char *filePath, uint32 charSize)
 {
     char fullFilePath[0x40];
-    sprintf_s(fullFilePath, (int32)sizeof(fullFilePath), "Data/Strings/%s", filePath);
+    sprintf_s(fullFilePath, sizeof(fullFilePath), "Data/Strings/%s", filePath);
 
     FileInfo info;
     InitFileInfo(&info);
